+++ lib/src/widgets/rotate.dart
@@ -1,19 +1,22 @@
 import 'package:flutter/material.dart';
+import 'dart:math' as math;
 
 class Rotate extends AnimatedWidget {
   final Widget child;
   final Curve? curve;
+  final double animationItensity;
 
   Rotate({
     required Animation<double> controller,
     required this.child,
+    required this.animationItensity,
     this.curve,
     Key? key,
   }) : super(
           key: key,
           listenable: Tween<double>(
             begin: 0,
-            end: 0.3,
+            end: 0.3 * animationItensity,
           ).animate(
             CurvedAnimation(
               parent: controller,
@@ -27,7 +30,7 @@ class Rotate extends AnimatedWidget {
   @override
   Widget build(BuildContext context) {
     return Transform.rotate(
-      angle: _translation.value,
+      angle: (_translation.value),
       child: child,
     );
   }