+++ lib/src/widgets/scale.dart
@@ -3,17 +3,19 @@ import 'package:flutter/material.dart';
 class Scale extends AnimatedWidget {
   final Widget child;
   final Curve? curve;
+  final double animationItensity;
 
   Scale({
     required Animation<double> controller,
     required this.child,
+    required this.animationItensity,
     this.curve,
     Key? key,
   }) : super(
           key: key,
           listenable: Tween<double>(
             begin: 1,
-            end: 1.4,
+            end: 1.4 * animationItensity,
           ).animate(
             CurvedAnimation(
               parent: controller,