+++ lib/src/widgets/shake.dart
@@ -4,17 +4,19 @@ class Shake extends AnimatedWidget {
   final AnimationController controller;
   final Curve? curve;
   final Widget child;
+  final double animationItensity;
 
   Shake({
     required this.controller,
     required this.child,
+    required this.animationItensity,
     this.curve,
     Key? key,
   }) : super(
           key: key,
           listenable: Tween<double>(
             begin: 1,
-            end: 7,
+            end: 7 * animationItensity,
           ).animate(
             CurvedAnimation(
               parent: controller,